requirements.txt
ccxt
pandas
numpy
optuna
matplotlib
pyarrow
pytest
python-dotenv
